+++ lua/custom/plugins/utils.lua
@@ -52,6 +52,20 @@ return {
     },
   },
 
+  -- [[experimental]]
+  -- The three "core" operations of add/delete/change can be done with the keymaps
+  -- ys{motion}{char}, ds{char}, and cs{target}{replacement}, respectively.
+  {
+    'kylechui/nvim-surround',
+    version = '^3.0.0', -- Use for stability; omit to use `main` branch for the latest features
+    event = 'VeryLazy',
+    config = function()
+      require('nvim-surround').setup {
+        -- Configuration here, or leave empty to use defaults
+      }
+    end,
+  },
+
   -- Detect tabstop and shiftwidth automatically
   'nmac427/guess-indent.nvim',