+++ lua/custom/plugins/utils.lua
@@ -67,7 +67,7 @@ return {
   },
 
   -- Detect tabstop and shiftwidth automatically
-  'nmac427/guess-indent.nvim',
+  { 'nmac427/guess-indent.nvim', opts = {} },
 
   -- translator
   'voldikss/vim-translator',